src/e/edit/JavaDoc.java: don't deliberately insert a line break between package and class names. It made sense when we were trying to stay narrow, but looks odd now.
src/e/edit/JavaResearcher.java: always show the non-private fields and static methods when showing the details of a class. (This makes up in part for the fact that we no longer offer JavaResearcher enough of the line to work out when we've typed "Class." or "new Class(", but we need to think about how to offer that kind of thing again in future.) Sort fields and methods alphabetically by name. Be more consistent with indentation (in the output, not the source).