universal.make: When the make-invoking code was moved into the package-for-distribution script, the comment was obsoleted but wasn't deleted or moved.
bin/package-for-distribution.rb: Said code probably does warrant a comment - someone will trip over this eventually.  I got bitten last time I used $(MAKE) (which was on this very line, before it was moved) because make -n thought that the rule should be run anyway.  That wouldn't happen here but I'm still loathe to make the change until it's needed and so testable.