bin/package-for-distribution.rb: Get the name of the make(1) program the user used out of the environment and use that, in case it's gmake (or has a path to the right version), rather than assuming that the "make" at the head of the PATH is OK.
universal.make: Make automatically exports MAKEFLAGS but not MAKE.  Interesting, but the manual doesn't seem to say why.  Perhaps because they only expect you to have to use it within makefiles.  Who would have make invoke a script and have the script then invoke make?  That'd be madness.  Ahem.

Astav Sacheti hit the make-3.80 "virtual memory exhausted" problem but was so sure that he was running make-3.81, as indeed he was, that he reported Terminator's -Xmx1g argument as the problem.