src/terminator/view/JTextBuffer.java: Never automatically scroll sideways on terminal output.  I haven't touched, or changed the behavior of, all the places in the code where we scroll to bottom via keypress or menu item.  That means that the original motivation for the auto-sideways scrolling code isn't broken by this change.  The original motivation for the scrolling was to scroll when editing long lines on a BlueArc Titan serial console - which doesn't auto-wrap like bash.  Editing involves keypresses - and they still scroll sideways.  We might want to rename scrollToBottom to make that explicit.
This is the first of a series of changes prompted by Min Xu's problem with mutt.  This one's only sufficient to fix my first simple demonstration of a problem:

The existing presence of a horizontal scrollbar is a necessary precondition to see the problem.  This line on its own causes no scrolling in a freshly created, default-sized window:

tput cup 40 79; echo -n a; tput cup 40 50

But when preceded by something to create a scrollbar of noticeable size:

ruby -e '(1..82).each() {|ii| $stdout.write("k"); }; puts();'

The tput line causes scrolling.