src/terminator/model/TextBuffer.java: My previous change didn't fix my occasional annoyance with aptitude update.  That turned out to be due to us allowing the cursor to be positioned off the right hand edge of the screen by using the horizontal cursor movement ("tput cuf") sequence.
gnome-terminal constraining the cursor movement in response to this:

tput cup 40 79; tput cuf 10; echo -n a; tput cup 40 40

That made me think that we should too - as we do for bidirectional cursor movement in the method above.  I'd noticed this inconsistency before but it's a pain to prove such things are wrong, rather than carefully thought out.

The next change will eliminate the need for this change, so even if this turns out to be flawed, we should still have solved today's problem.