src/terminator/view/JTerminalPane.java: Since I earlier stopped us scrolling horizontally when we're at the "sticky bottom" of the scrollbar, the scroll-on-keypress is all that's attempting to keep the cursor visible.  Scrolling to the cursor as soon as the key's pressed - like we were doing before this delaying patch - only works when the key causes no display output.  When the key causes display output, that output will follow somewhat later.
If you're to have a good interactive experience, the output has to happen in roughly human reaction time which is ~100ms.  That's usually the case when we're typing into cat(1) or Titan serial consoles.  There is fundamentally no way to reliably tie up keyboard input with remotely generated output.  Assuming that anything which happens within one reaction period of a keypress was the direct consequence of that keypress seems about the best we could do.

If we don't do this, the automatic horizontal scrolling works one character behind in the normal case.  That's really quite tedious.