src/e/edit/JavaDoc.java: fix the HREF in our generated links to source.
src/e/edit/Evergreen.java: handle non-HTML URLs ourselves.

src/e/edit/Advisor.java: arrange for HTML to be displayed in the documentation browser, but for links to source to be passed to the editor. This fixes the "[src]" links in our Java documentation, though I wonder if we shouldn't have gone with a made-up scheme such as "x-src:" instead, for clarity?