bin/invoke-java.rb: If the first line of the InAppServer's output is an authentication status message, act appropriately - removing it silently on success and return failure otherwise.
Many people won't see the warning message I added a few minutes ago to $stdout because they start new Terminator windows by clicking an icon or a button.  So this change makes Terminator fall-back to opening a new process.

I *think* it should be benign when talking to older versions of InAppServer.java.