src/e/ptextarea/PMakefileTextStyler.java: $(lastword) was the motivation for refreshing this list from http://www.gnu.org/software/make/manual/html_node/Name-Index.html which is now from the make-3.81 tree.  I used this to filter-out the functions:
ruby -ne '$_.match(/<code>([a-z][^A-Z]*)<\/code>/) && puts("\"#{$1}\",")' Name-Index.html | sort -u

And then manually edited out "make" and "makefile".

(Yes, $(and), $(or) and $(info) are new functions, although this is the first I've heard of them too.  $(notdir) was, however, never anything to do with Reading's Inner Distribution Road.)