src/e/forms/FormDialog.java: Defer to the constraining code in JFrameUtilities, which understands insets and the like.  Make sure we pass it the previous location and size, by setting the location and size of the dialog first.
src/e/gui/JFrameUtilities.java: The user would usually rather have a moved window than a squashed one, so try moving the window before shrinking it.  The previous code didn't cope well with windows where the top left corner was off the bottom right of the screen.  That seems reasonably likely to happen when you change resolution.  I've effectively fixed that, along with the comments (which suggested that it was already OK).

This moves Open Quickly and Find in Files so that they're on-screen and minimally squashed.