src/e/ptextarea/PTextArea.java: take PTextArea out of the tab-insertion loop...
src/e/ptextarea/PKeyHandler.java: ...and make shift-tab insert a hard tab, regardless of what the indentation string is. I've been meaning to implement some way of easily entering hard tabs for getting on for ten years now. I'd assumed it would be on control-tab, but that's since become popular as a "cycle through the tabs in a tabbed pane" keystroke.