src/terminator/view/JTerminalPane.java: Min Xu's still not happy - he's still having problems with mutt and vimdiff.  I did reproduce his problems last time but only intermittently.  So I came up with a one line piece of Ruby test code (which I've left in a comment) to simulate what might be happening when you press the down arrow when vimdiff's on the end-of-line position on a file with long lines.
Cancelling the timer when starting it helped and upping the timeout to 200ms helped when using cat on jessies.org from California.  But the majority of the change here is to check that output has quiesced before deciding to scroll.  Min Xu's previously opined that the mutt and vimdiff cursor is only off-screen briefly.  This stops that sort of behavior from causing scrolling and, indeed, stops automatic scrolling if output is continuing.