src/e/ptextarea/PKeyHandler.java: martind noticed that you could enter newlines and tabs in a PTextArea on which setEditable(false) had been called. I wonder if we shouldn't move the "editable" property down into the text buffer instead, where it would be easier to police? We'd lose the ability to have two views with different editable states for one model, but is that useful? We'd still need to keep our current isEditable checks anyway, though, to avoid weird effects like moving the caret/changing the selection despite not actually having changed the text (because the edit was ignored at a low level). And it's important to remember that the semantics of setEditable(false) are that programmatic edits are still allowed. Applications would have to add a bunch of awkward and error-prone code if we were to change that.
An alternative would be to say "you can't use the non-modifying keys either" and just block all key input, but that would mean we'd lose the ability to select and copy text from the keyboard, and that's too useful to lose.

All in all, I think the current situation isn't easily improved. The problems fixed by this patch were introduced fairly recently, and we haven't had much trouble with this class of error. (And how many more keys can we have left to add/fix support for?)