Nikos Tzanos isn't the first person to trip over "the ubygems problem" when trying to start Terminator.  (He was certainly following in my footsteps.)
www/index.html: So we should mention it on a web page.  The potential problems when starting our shortcuts are mentioned on the Terminator web page, even when they'd apply to Evergreen too, so I've done that here too.

TODO: We don't need two copies of this information about "ubygems" but the discussion about what we need to do to support the native Ruby probably does still belong in TODO.