www/header.html: put the TABLE for the links at the top of the page in a DIV so we can set the background color of the DIV to get pretty much the effect I'd originally wanted. Yes; this is how modest my ambitions were.
www/styles.css: use the DIV to add color, rather than the TABLE. Switch to "position" rather than "margin", because that's what the web seems to think we should use. This still isn't saying what I want though, which would be something like "for every H\d, indent by 10px; similar for every P 'within' them, and a little more for every PRE". One problem is that the containment is wrong: H1 only contains the header text, not all the content under that heading. To do the right thing, I think we need to dump the built-in headings stuff and just use DIVs with custom classes. But for now, this is close enough, with the proviso that we can no longer have text that isn't part of a P. (I don't know what the technical term for this is. If I did, I might be able to work out the CSS selector for it, and remove the restriction.)

www/index.html: for now, we'll have to add the missing P tags. While I'm here, update the Microsoft link (the old page said "this is no longer current; use [new link target] instead").