src/e/edit/Workspace.java: I believed that there was zero-or-one FindInFilesDialog and zero-or-one OpenQuicklyDialog for each Workspace. I was right about the OpenQuicklyDialog. This bug meant that the longer you used today's build, the more background searches you'd be collecting, and the slower it would get until you had to restart.
src/e/edit/FindInFilesDialog.java: but first I wandered down a bit of a blind alley trying to work out what was broken about FindInFilesDialog itself. In part because of something martind had complained about, which was the serialization of find requests, where you couldn't (for example) run multiple finds concurrently. This hid today's real problem, and I think that the separate (and per-Evergreen) thread pool for definition finding is sufficient protection against bringing the user's machine to its knees. So remove that restriction, and tidy some of this code up a bit.