src/terminator/Options.java: martind points out that with the last revision:
> if I uncheck the block cursor setting,
> then it disappears from ~/.terminator/options but remains in
> ~/.terminator-settings.  This means that block cursor setting comes  
> back when I start a new Terminator.

so let's try more aggressively to get rid of the old file. We'll still read in the old file, but then we'll try to write it out in the new location, and if that was successful we'll try to remove the old file. This should move everyone [who upgrades] over as quickly as possible.