universal.make: Include the version number in the .deb filename, like Debian do.  This might make Ubuntu's software update and the like happier.  It seems likely that, if you have a cached, out-of-date version of the Packages index, things are less likely to get upset if the file you download is the one you're expecting rather than a newer version.  This has the added advantage that we'll leave older .debs around for a while, like we were already doing for .rpms...
www/downloads.html: ... and the (arguable) disadvantage that the .rpm name will have to change to include the architecture, if we're to keep things simple.  The flipside is that we're likely to support x86_64 rpms as soon as we support amd64 .debs.

I'll need to do some invisible work to rename or regenerate the rpms.