bin/ithaki.sh: Undo the hacks now I've (a) fixed libx11-dev on wide.us.dev.bluearc.com and (b) fixed the makerules etc to cope with different architectures on the same OS sharing the same work area.  (It didn't require a very big hammer after all - apt-get was just asking me, in a particularly obtuse way, to ask it to install that package individually, so it could tell me that it would have to remove some obsolete package that one of its dependencies replaces.)
www/downloads.html: Flip the links to the new downloads/ tree, whose one .htaccess file is now conveniently under source code control.  Advertise the amd64 builds.  I haven't seen how this looks in a browser.  I'm not attached to the form of my changes - I was just trying to minimize the amount of duplication in the HTML to minimize the likelihood of my making a mistake.