bin/terminator: Interesting that no-one's noticed that this call to finish the gnome startup nonsense has been broken since:
2006-10-22 18:45:12 -0700 enh (1813)

I still wouldn't have noticed except that I've "broken it more" by renaming the Linux directory and I wouldn't have noticed that except that I was searching for the gnome-startup stuff because some other code is now trying to run gnome-startup (with the "start" option) on Cygwin.

I haven't tested this, so it probably still doesn't work.  Looking at what happens when I start a second Terminator, is this code obsolete anyway?