src/e/edit/WorkspaceFileList.java: The isSymbolicLink call had been inside the check for isDirectory since isSymbolicLink bogusly returned true for files with symbolic links on their paths.  If the isSymbolicLink had been where I've now put it and your workspace was underneath a symbolic link (as many of mine have historically been) then the file list would have been entirely empty.  Shortly after the Subversion import, however, isSymbolicLink's bogosity tripped me up in another way and I came up with a (blog-worthy?) fix.
Now I've just tripped over the call being in the wrong place.  For reasons possibly due to a temporary hack I put in BlueArc's makerules, I have the following symlink:

/home/martind/work/parrot/libs/eventlog/generated/new/generated -> .

I'm surprised that Evergreen coped gracefully with that.  I saw no exceptions and had no cause all day to think that the file list had been truncated (so maybe it was and maybe it wasn't).  The file list did, however, include some very long file names.

I have explicitly tested this using a workspace whose root directory is a symbolic link.  The first time we apply the isSymbolicLink test is on the contents of the root directory, so we never check the file type of the workspace root itself.