universal.make: Remove all the @ prefix from the start of (almost) all rules.  That hides the commands we're executing, which is unhelpful in the case of error.  "(almost)" because there are a number of exceptions:
(1) When we compile Java, don't echo the whole, enormous command.  Java compilation is not done in The Make Way, so it's not really an extra special case, just a continuation of an existing one.  It would be more useful to echo how we calculate the list of sources.

(2) When we echo something in the rules, there's no point in echoing the echo command.

(3) When we're recursively invoked to create the installer's file list, our output is parsed and so mustn't contain the command (or the parser would have to cope).

Also:

* Lose the preparatory comments, mainly to simple.make (though I've rewritten the ones about the location of variables and moved them later in the file).

* The BlueArc makerules maintainers eventually decided that the behavior of .SECONDARY wasn't what we wanted.  It meant that a source file which wouldn't compile wouldn't break the build providing that its executable was newer than the executable's extant dependencies.  This didn't cause me a problem here but I thought it was causing me a problem.

* In the spirit of not saying "the world" when we mean "the Java", rename SOURCE_FILES.

* Lose the careful echo -n changes from the previous check-in.  Our output is going to be much more terse than it has previously been in the normal case but more verbose in the case of error.  With the -n change, we'd look weirdly untidy in the normal case or we'd compromise our completeness in the case of error (or we'd have to do something more complicated in filter-build-output.rb).

per-directory.make: As with universal.make, switch to echoing the commands normally.  Here might be a reasonable place to mention that I was recently disabused of my long-held superstition that you have to join all commands in a rule with "&& \" to ensure that later commands aren't run in the case of earlier errors.  I'm not sure where that superstition came from but it doesn't seem to be true in any circumstances I can reproduce.

simple.make: A new pain in the butt, which exists solely to change the user's invocation of "make" into make 2>&1 | filter-build-output.rb.  I was going to make this the new universal.make but Subversion's support for renaming wouldn't let me then look at the whole history of what has historically been called universal.make.  I'd rather pay the price of changing one line in all the other repositories.

bin/filter-build-output.rb: Filter the output from universal.make, so we can produce all of the possibly relevant output in the case of an error, while still providing an indication of progress in the normal case.  This is the kind of output you see with this version:

martind@duezer:~/software.jessies.org/work/salma-hayek$ make
Compiling Java source...
Compiling gnome-startup.cpp...
Compiling launcher.cpp...
Compiling synchronizeWindowsEnvironment.cpp...
Compiling setsid.cpp...
martind@duezer:~/software.jessies.org/work/salma-hayek$ 

For one thing, this should make it a lot easier to spot errors in the nightly build mails.

The Mac OS "disk image" creation rule/filtering might want some work.  The WiX rules could use some work to echo what they're doing.

Makefile: The external make interface is now called simple.make.