bin/package-for-distribution.rb: make installer-file-list produces no output any more because it gets filtered out by filter-build-output.rb.  We'll need to use universal.make directly.
That's easy enough but my first effort didn't work.  I had trouble debugging it because the error message wasn't shown.  The error message wasn't shown because the "Building .deb package" message came out after it, causing filter-build-output.rb to conclude that all previous output was from successful commands.  A flush cures that.  You could argue that the "Building .deb package" line is now spurious but I don't think it is - this is still a big, complicated step in the installer rules - but the message could be improved, given that it's not this code which actually generates the .deb.

bin/filter-build-output.rb: It wasn't my intention to stop pumping errors at the first one.  That wasn't the cause of my not being able to debug my change to package-for-distribution but I thought it might be.