bin/mdorey.sh: When svn tried to update mdorey.sh it got a permission denied error because it was already open by bash.  Subversion then left my work area in a broken state which I seemingly had no option but to manually repair by removing bin/filter-build-output.rb (which still left svn status saying that . and bin had status "!") followed by svn update which seemed to sort things out.  Perhaps catting the file into bash would let it be updated.
I don't have this problem in ithaki.sh because there I run the script from my main work area - to ensure it's always up-to-date (with my changes).  I can't do that from my Cygwin machine because the way cron is set up, it can't access the network.  The alternative is that it breaks when I'm not logged in (as when simonj and I swapped offices for a week).