src/e/edit/Evergreen.java: speed start-up by only opening those files that are going to be visible. Basically, we keep a list of workspaces-to-be, each of which has a list of files-to-be. We create all the workspaces early on, because we need them to populate the tabbed pane. But only as a workspace gets selected do we actually load its files.
src/e/edit/Workspace.java: this class is now responsible for setting the tabbed pane tab titles, add has to add the number of actually loaded files to the number of pending files to work out what number goes in parentheses in the title.

src/e/edit/EColumn.java: this class now tells the workspace to update the tab title, rather than doing it itself. We also have to make sure we set the width of newly-added components; we always did in the heuristic case (which is what you'd use if you were adding files after the application has started) but used to get away without doing it in the non-heuristic case, for some reason, and now we don't. Perhaps because before we were added before the tab was first shown (and maybe it stretched us to fit when it was shown?) but now we're added after it's already showing?

src/e/edit/JavaResearcher.java: make it easy for the start-up code to spin initialization of this class off onto a separate thread.