Switch to using the packaging script to generate the installer tree for the Windows installations, giving us a (hopefully) quicker startup time through the use of classes.jar.
universal.make: There were a couple of steps:

Remove support for the WiX module format that we used for salma-hayek.msm - the packaging script packages a copy of salma-hayek inside each installed project.  This means that salma-hayek's native-dist target no longer builds an installer even on Cygwin.

Use the packaging directory's file list to produce the list of files to include in the WiX XML.  That means we need to cd into the packaging directory before calling light(1).

lib/installer.wxs: The packaging script makes a better job of the documentation than I was doing here.  We no longer want to pull in a installer "module" for salma-hayek.  The directory structure is now fairly different, which mainly affects the desktop shortcut - the working directory and the path to the icon need to change.

bin/package-for-distribution.rb: Package the documentation for Cygwin as for the Mac.  jar(1) is a Java program which uses "JVM-compatible" paths.  This caught me out because previously we weren't saying which exactly command we were executing here and, more seriously, we weren't checking for failures.  Ruby's system method behaves in all manner of different ways according to how you call it.  Wrap the variant we use to call jar with some better error checking and give it a better name.