I wanted to experiment with a modified indentation guesser.  To do that a bit more scientifically than usual, I wanted to test how many lines the indenter would change, given the current indentation guesser's algorithm.
src/e/demo/PTextAreaDemo.java: So I wrote countLinesChangedByIndenter.  It failed with an out-of-bounds exception due to a bug in PLogicalSegmentIterator.  Even after the fix (below), I didn't want to call it on every invocation of PTextAreaDemo because it mutates the document (and, probably because it does so before the document is showing, it messes up various things related to SplitLine).  iterateOverSegments is sufficient to demonstrate the problem and quick enough (<100ms for PTextAreaDemo.java, ~1.1s for a 1MiB log file) that we can do it on every invocation to protect ourselves from regression.

src/e/ptextarea/PLogicalSegmentIterator.java: Don't add a newline segment after the last character in the buffer.  We can't do charAt(0) on that segment, despite its length being 1, and the toString method similarly throws an exception.  Rewrite the duplicated code in the constructor to avoid duplicating the fix (and to be smaller and simpler).

I tested by asking for segments starting beyond the last character as well as checking that the segmentCount is one lower now than it was.