lib/installer.wxs: Finally the light dawned.  The way WiX seems to want me to say ifdef(RUBYW_HKLM) ? HKLM : HKCU is to duplicate everything that depends on it then install one or other of the duplicates in its own "Feature".  This worked on my Windows box and JohnH's.  I have an all-users installation and he has a just-me installation.  We're both administrators, so HKMU binds to HKLM, given that I set ALLUSERS to 2.  Hopefully JohnH's installation is the worst case - where an administrator chose a just-me installation for Cygwin.  He didn't originally have Ruby installed, so the new version of that condition was tested too.
I didn't start out to fix this.  I was trying to see if WiX 3 would give me any new tools for solving it.  It does have some predefined stuff which might make it easy to satisfy Aryeh Leib Taurog's request for a where-should-we-install UI.  But I didn't get as far as downloading it before I realized the solution above.