src/e/edit/ETextWindow.java: update the tags as a side-effect of fillWithContent, the specific bug I'm fixing being that "Revert to Saved" wasn't causing the tags to update.
src/e/edit/TagsUpdater.java: we were updating tagsDigest at a point where we might still throw away our newly-built tags because, when it came to show them, we didn't have the latest sequence number. But if the thread with the latest sequence number checked the digest after we'd updated it, it would decide that the tags hadn't changed, because tagsDigest no longer corresponds with the displayed tags. So we must update tagsDigest only if we update the displayed tags. This was probably the root cause of many of the problems we've been seeing lately with tags not getting updated.