src/terminator/view/JTerminalPane.java: Make it less likely that any reader will assume we're out-of-date, instead of finding that they're actually using an out-of-date terminfo or an out-of-date xterm.  Elliott's post-check-in clarification was:
> amusingly, though Terminal.app actually uses the modern XTerm  
> sequences, its terminfo specifies the old ones. that is, Mac OS has  
> outdated XTerm infocmp. which i guess is more proof that  
> applications do as XTerm does rather than as infocmp would tell them.

actually, it's stranger than that. on my Ubuntu box, xterm-color has  
the "old" values for F1-F4 while "xterm" has the new ones. but both  
have out-of-date values for the cursor keys. so although infocmp(1)  
says we differ from XTerm, we don't. that's a big extra gotcha for  
trying to match XTerm; i'd assumed that we just had to work until the  
output of "infocmp -1 xterm terminator" was as close to empty as  
possible, but i don't think that would be wise.

"ctlseqs", on the other hand, does seem to be kept up to date. i  
haven't found a discrepancy there yet. but it's probably going to be  
wise to always actually test XTerm.

i also notice that although Ubuntu 6.10 is on XTerm(216), dickey's up  
to XTerm(224) 2007-02-11, so it's still a moving target.