src/e/util/ProcessUtilities.java: add support for working out the current JVM's pid on Mac OS as well as Linux.
src/e/gui/DebugMenu.java: disable the jmap(1) histogram on Java 5, where it doesn't work properly. I think we could fix it by not displaying the results in the JVM that's being examined, but since later versions of jmap(1) don't produce any output until they've finished, we get away with it (I don't know that this is the problem, but sample(1) showed jmap(1) seemingly stuck in Java_java_io_FileOutputStream_writeBytes, so it's a distinct possibility).