native/all/launcher/launcher.cpp: Kevin Martin helpfully included some ls output with his problem report:
$ ls /proc/registry/HKEY_LOCAL_MACHINE/SOFTWARE/JavaSoft/Java\ Runtime\ Environment/
1.4  1.4.2_10  CurrentVersion

That made diagnosis very straightforward.  We can automate that - rather than throwing the information away, which we were previously doing.

I changed the commented-out 1.6 test to be against 1.7 because commenting it out got harder.  We could just delete it.  It wouldn't be rocket science if we wanted to reinstate it.