bin/invoke-java.rb: use a technique mentioned long ago on the ruby-talk list (and seemingly ignored, despite being a great idea for rubyw.exe) to effectively add an uncaught exception handler. There doesn't seem to be a way to do this globally in Ruby, so we might want to make some kind of abstract Invoker class and let complicated scripts such as bin/terminator provide behavior and have it run by the invoke method.
bin/show-alert.rb: rewrite to make it clear that you get one show_alert function which has the same signature on each platform. It was only ever inside-out because the three were written independently. Ditch the Applescript, because it's slow, ugly, unreadable, unmaintainable, and doesn't play nicely with escaping/quoting. I fought for long enough trying to get it to behave before giving up and writing...

native/Darwin/NSRunAlertPanel/NSRunAlertPanel.mm: a trivial utility to call NSRunInformationalAlertPanel. I couldn't work out how to do this using Ruby's "dl/import". We should be able to use RubyCocoa when we switch to requiring Java 6 (which isn't likely to ever be available for Mac OS 10.4). This code doesn't get the right icon, but neither did the Applescript. It doesn't seem worth fixing this; we can fix the RubyCocoa version when we have it. (Plus no-one should ever see this anyway.)