src/terminator/model/StyledText.java: we were previously quite sloppy about the foreground and background values for styles that didn't have a specific foreground or background. The value would depend on how we'd arrived at the present style, meaning the "short" representation of a style wasn't suitable for comparison. That didn't matter before (as far as I can think), but it matters now...
src/terminator/model/TextLine.java: rewrite the style-handling side of this class to lazily create the short[] of style data the first time non-default styling is needed for that line. The reasoning being that most lines only ever use the default style, and many of those lines that do use non-default styles are short-lived (because they're on the alternate buffer because you're running something like Vim). Before, the cost per character was two bytes for the character data plus two bytes for the style data for each character (plus some per-line overhead). Now, the cost per character is two bytes for the character itself, and you only pay for the extra two bytes of style data if needed. There's some unnecessary copying in the new code, both for simplicity's sake (copying with the case where there's no style data has added some complexity) but also because with the likely size of a styled line (<= 80 characters) it's not obvious that avoiding copying a few bytes saves more than the extra code and method calls cost; we should measure, I think, before we rush to put the cleverness back.