src/e/ptextarea/PTextArea.java: Fix the regression introduced with the LazyStringSelection change whereby clicking to position the caret in the target text area would clear the X11 selection, preventing middle button pasting.  copyToClipboard, which we no longer call, contains some extra logic (not implied by its name) to avoid touching the clipboard when the selection is empty.  Without this (now reintroduced) check, I assert that middle button pasting is significantly less useful.
The corresponding code in Terminator doesn't need fixing so urgently because SelectionHighlighter.selectionChanged isn't called if the selection becomes empty.  So you can click on a Terminator window (so clearing that window's selection) and still middle button paste to insert text from elsewhere.  What you can't do is select some text, click in the same window and paste... unless you also pasted before clicking, in which case the selection has been cached.  Eugh - that's a nasty implementation detail to be exposed.  It's less common to make large selections in a terminal than an editor and still less common for them to be for purposes other than pasting.  Perhaps we should back out the laziness from Terminator.

Even with this fix, PTextArea's behavior has still changed from how it was before the LazyStringSelection change.  It used to be that you could make a selection, position the caret elsewhere and middle button paste the selection.  Now you can't do that... unless you cause the selection to be cached by pasting it over the top of itself before repositioning the caret.  Eugh.  So perhaps we should back out the laziness from PTextArea too.

gedit pastes wherever the mouse pointer is when you press the middle button.