bin/ithaki.sh: Collapse the duplication to improve clarity...
bin/mdorey.sh: ... and switch to the same form of build tree here, running the scripts from a "work" area but building in a virgin "nightlies" tree.  The reason for the former is so that the builds aren't made from yesterday's scripts, something that often causes problems because it's easy to check in changes which require the latest build scripts.  (At first thought, it looks better to have the script update the nightlies area and then run a second-stage script.  That's a bit tricky because both Ruby and bash rely on the script not changing under the interpreter's feet.  Perl doesn't.  bkmakeall uses a different technique, involving copying itself to /tmp/.  None of those solutions seem worth the effort and I haven't suffered any pain from my current approach.)

My previous change worked - so we have a Windows build from today.  I'm just taking advantage of the improved nightly-build.rb to do things the way I've found to work well for Linux.