bin/ensure-suitable-mac-os-version.rb: don't worry about Java here; just worry about the Mac OS version, as the script's name suggests. This script predates me making the Java-version-checking code available for all OSes, so removing this lets us take advantage of the (since) more sophisticated code in "invoke-java.rb". The only drawback is that we now don't tell the user all of their problems at once, but practically speaking that's unlikely ever to be a real problem because Apple Java releases tend to be tied to OS releases, and if they weren't, you'd only have one thing to upgrade anyway. "package-for-distribution.rb" ensures that this Mac OS version test is run before our Java version test, and that's the important thing.
bin/package-for-distribution.rb: stop messing with the path. It's now fully "invoke-java.rb"'s problem to worry about finding a suitable java(1). This will have the effect of breaking any Java (not ours) that was (unknowingly) relying on being run by our Java to work around an unsuitable java(1), but that's the user's (or other author's) concern, not ours.