universal.make: Chris had a problem connecting with ssh to software@jessies.org in these rules.  He noted that we were trying to symlink to a package that didn't exist.  Looking at the rules, that wasn't part of his problem - we always do that - but it's confusing and opens a window in which a failure will leave us with a dangling symlink for the latest version of a package.  So I've reversed the dependency between the upload and symlink-latest rules.
I've looked at the history and this dates from when only the Linux packages' names contained the version numbers.  I hook the symlink-latest rule into the dependency chain with a little hack, to keep the code small and isolated.  I forgot to signpost it and so forgot to revisit that when I decided that the version numbers would be useful on the other platforms too (in order to make it easier to keep the last few versions around).

I'm not sure why I implemented the dependency for INSTALLERS rather than STANDALONE_INSTALLERS.  Now there's no longer a WiX "installer" for salma-hayek, I'm not sure there's any value in the distinction but, if I'm keeping it, I should use it consistently.  And if I'm getting rid of it, that should be a separate check-in.