universal.make: We've never been able to work out when we don't need to recompile the Java.  The problem comes about because we don't know which .class files are generated.  One .java file can generate several .class files.  If one of the inner classes' .class files disappears, how will we know to rebuild?  This could plausibly happen if someone hits Ctrl-C during the compilation (or we run out of disk space or...).
We can guard against incomplete compilations by using a sentinel file - one that's touched after a successful compilation.  If some mechanism then causes an inner class's .class file to be deleted, that will touch its directory timestamp.  If one of the subdirectories of classes/ is removed wholesale, that will touch its parent directory's timestamp.

The Java also needs recompiling if any source files have been added or deleted.  Added source files are easy and deleted ones we can handle with directory dependencies, as proposed above for .class files.  We need to jump through one hoop to stop rebuilds on every svn update.  svn update touches its .svn directories and I don't think we care about those.

This all smacks of thinking I'm smarter today than I was before.  I've (only) found one problem so far - if I remove classes/ (but not .generated/), then classes/ doesn't get regenerated.  That can wait for the next check-in.

per-directory.make: The rule for generating the JNI .class file on a dependency of all the Java source was a pretty safe way of speeding up "make native" for Terminator, by removing the need for it to recompile the Java every time.  It wasn't quite safe, in that removing a .java file or a .class file wouldn't cause a recompilation, but that never caused a problem in practice.  If the sentinel works, though, we can have complete safety and still not pay for recompilation every time, as we would have, were we to have depended on the old build.java.

Note that this doesn't cause the JNI C++ to recompile every time the Java is recompiled.  We generate a new JNI header every time the Java is recompiled, as we should, but we only overwrite the copy we compile if it's changed.