universal.make: Support a new "test" target, in almost exactly the same form as I suggested to Phil, off the cuff.
bin/filter-build-output.rb: Relax the regular expression so that my "Testing..." message comes out.  That's all that comes out unless run_tests returns a failure, in which case we'll see all the output...

tests/run_tests: ... so we should return failure if we see anything unexpected.  At the moment, I only do this if we see an unexpected failure on any test.  We should consider returning failure for an unexpected success, which would indicate that the test needs changing.

simple.make: "Publish" the new target to any human reader.